Cleaned up the code a little and fixed Linux and Mac paths to not point to CK2

+++ src/Main.java
@@ -6,11 +6,11 @@ public class Main {
 		String[] oss = {"Windows", "Linux", "Mac"};
 		JComboBox<String> osCB = new JComboBox<>(oss);
         JOptionPane.showMessageDialog(null, osCB, "Choose the operating system:", JOptionPane.PLAIN_MESSAGE);
-        int os = osCB.getSelectedIndex();
+        byte os = (byte) osCB.getSelectedIndex();
 
         JComboBox<String> gameCB = new JComboBox<>(Mod.games);
         JOptionPane.showMessageDialog(null, gameCB, "Choose the game", JOptionPane.PLAIN_MESSAGE);
-        int game = gameCB.getSelectedIndex();
+        byte game = (byte) gameCB.getSelectedIndex();
 
         new MyFrame(os, game);
     }
